% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_cell.R
\name{meta_cell}
\alias{meta_cell}
\title{Select meta cells from huge single cell dataset}
\usage{
meta_cell(sc_data, cell_types)
}
\arguments{
\item{sc_data}{scRNA-seq data of the original large dataset}

\item{cell_type}{Clusters, each each cluster, the most representative cells will be selected}
}
\value{
selected meta cells
}
\description{
Select meta cells from huge single cell dataset
}
\examples{
meta_cell(sc_dataset, Clusters)
}
