% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_python.R
\name{SCTP}
\alias{SCTP}
\title{Run deep learning training process with python torch module}
\usage{
SCTP(sc_expr, cell_types, st_dataset, bulk_dataset, pheno, out_dir)
}
\arguments{
\item{sc_expr}{scRNA-seq data}

\item{cell_types}{Clusters, each each cluster, the most representative cells will be selected}

\item{st_dataset}{spatial transcriptomic data in matrix}

\item{bulk_dataset}{bulk gene expression matrix}

\item{pheno}{phenotype of bulk samples}

\item{out_dir}{the path in which the features and networks will be stored}
}
\value{
estimated coefficient and the scRNA-seq and ST data used in the model
}
\description{
Run deep learning training process with python torch module
}
\examples{
SCTP(sc_expr=sc_filtered, cell_types=filtered_annot$Cell_subtype, st_dataset = st_dataset, bulk_dataset = bulk_dataset, pheno=pheno, out_dir="OUTPUT")
}
