% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCTP_prepare.R
\name{SCTP_prepare}
\alias{SCTP_prepare}
\title{Run deep learning training process with python torch module}
\usage{
SCTP_prepare(sc_seurat, st_seurat, bulk_dataset, pheno, out_dir)
}
\arguments{
\item{sc_seurat}{seurat object of scRNA-seq data}

\item{st_seurat}{seurat object of ST data}

\item{bulk_dataset}{bulk gene expression matrix}

\item{pheno}{phenotype of bulk samples}

\item{out_dir}{the path in which the features and networks will be stored}

\item{st_dataset}{spatial transcriptomic data in matrix}
}
\value{
create networks and features and phenotype for model training
}
\description{
Run deep learning training process with python torch module
}
\examples{
SCTP_prepare(sc_expr=sc_dataset, st_dataset = st_dataset, bulk_dataset = bulk_dataset, pheno=pheno, out_dir="OUTPUT")
}
